Introduce StepFilterProvider
